backend/requirements.txt
pydantic
sqlalchemy
psycopg2-binary
passlib
bcrypt==3.2.2
python-jose
joblib
scikit-learn==1.6.1
pandas
numpy
fastapi
uvicorn[standard]
Jinja2
python-multipart
python-dotenv
paho-mqtt
# tensorflow==2.19.0
tensorflow-cpu==2.19.0
jinja2
bcrypt==3.2.2
